+++ CHANGELOG.md
@@ -4,6 +4,10 @@ This file documents all notable changes to the GEOS-Chem repository starting in 
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased] - TBD
+### Fixed
+- Fixed Hg directional ocean flux diagnostics in the Hg simulation so that they equal net flux
+
 ## [14.6.3] - 2025-07-28
 ### Added
 - Added error check to exclude sampling ObsPack observations located outside of a nested-grid domain